move multi value creation logic to CreateResourceV2Handler
